renderer optimized could lead to issue -> remember

+++ src/main/java/ch/g_7/graphite/base/entity/object/BasicGameObject.java
@@ -57,7 +57,7 @@ public class BasicGameObject implements IGameObject{
 	}
 
 	@Override
-	public double getScale() {
+	public float getScale() {
 		return scale;
 	}
